+++ packages/fast-figma-plugin-msft/legacy/interface/strings.ts
