+++ packages/fast-figma-plugin-msft/legacy/utilities/design-system.ts
