+++ packages/fast-figma-plugin-msft/legacy/utilities/paint.ts
