+++ packages/fast-figma-plugin-msft/src/ui.tsx
@@ -1,17 +1,29 @@
-/**
- * to communicate with the parent script.
- */
 import React from "react";
 import ReactDOM from "react-dom";
-import { PluginUI } from "./interface/plugin-ui";
+import { PluginUI, PluginUIProps } from "./core/ui";
+
+function render(props?: PluginUIProps): void {
+    console.log("RENDER", props);
+    ReactDOM.render(<PluginUI {...props} />, root);
+}
 
 // Import with require so the dependency doesn't get tree-shaken
 // tslint:disable-next-line
 const styles = require("./global.css");
 
+/**
+ * Create root element
+ */
 const root: HTMLDivElement = document.createElement("div");
 root.setAttribute("id", "root");
 document.body.appendChild(root);
 
+/**
+ * Wire iframe's onmessage to render function
+ */
+window.onmessage = (e: any): void => {
+    render(e.data.pluginMessage);
+};
+
 // Render UI
-ReactDOM.render(<PluginUI />, root);
+render();
